rename event_list to event_linked_list
